Add files via upload
New install.json in .tar.gz file.

WHM 58 has an error with processing name strings where if it contains an apostrophe, it doesn't process the string and results in $LANG{'Let'sEncrypt'} being displayed in the customer cPanel. WHM 56 did not have this issue.

Changing install.json name attribute to use the HTML code for this name resolves the issue;
[
        {
                "icon":"letsencrypt.png",
                "featuremanager":true,
                "name":"Let&#39;sEncrypt",
                "uri":"/3rdparty/letsencrypt-cpanel-ui/index.live.pl",
                "group_id":"security",
                "order":"9",
                "id":"letsencrypt",
                "type":"link"
        }
]

No other changes made to the .tar.gz file or it's contents. Tested and confirmed working on our production server.

-KP